fix: renamed getUserApi to userApi
